Fix CI failures, lint all targets, stabilize the browser (#186)
* fix(tests,ci): throttle oasis-app e2e suites and de-flake two wall-clock asserts

CI run 35814809095 failed only on timing on a loaded nextest pool:

- e2e_input::key_flows_on_every_skin and
  e2e_shell::icon_launch_and_close_on_every_skin were killed at the
  5-minute guard (~30-40s when not starved). Put the oasis-app
  integration tests in a nextest test group capped at 8 concurrent
  tests and give them a 10-minute hang guard; the Test step timeout
  still bounds the run.
- e2e_browser::closing_the_browser_during_a_slow_load_is_prompt_and_clean
  read 955ms against a 500ms bound. Stretch /slow to 6s and bound the
  close at half of that, so a close that joined the I/O thread is still
  unmistakable.
- e2e_user_flows::radio_tune_offline_reports_an_error_and_favorites_persist
  bounded 30 frames at 5s total. Bound the slowest single frame at 3s
  instead: a blocking connect stalls one frame, scheduler load spreads
  over all of them.


* fix(tests): make the StreamingBuffer Send + Sync assertion const-evaluated

The compile-time check wrapped the call in an uncalled `fn check()`,
so rustc reported both helpers as dead code when building oasis-app's
tests. Make the helper a `const fn` and call it directly in the
`const _` block: still a compile-time assertion, nothing left unused,
no `#[allow]` needed.


* ci: lint tests and benches with clippy (--all-targets) and fix the fallout

CI ran `cargo clippy --workspace` only over lib/bin targets, so test,
bench and example code was never linted. `--all-targets` reported ~4.5k
errors. Fixes:

- clippy.toml: `allow-unwrap-in-tests = true`. `unwrap_used` stays denied
  in shipped code; in tests a panic on Err/None is the failure report.
  This covered ~4.1k of the errors, all of the oasis-js ones included.
- field_reassign_with_default (~150 test setups): rewritten to
  struct-update syntax (`T { a, ..Default::default() }`).
- Tautological / constant assertions: the blur filter test's
  `x < y || x >= y` checks now assert the colour really desaturates;
  checks on constants became `const { assert!(..) }`.
- tv_controller tests used `read()` where a short read would pass
  silently; they now use `read_exact()`.
- Machine-applicable lints via `clippy --fix` (while-let loops,
  collapsible ifs, manual range/div_ceil, test modules moved to file end,
  etc.). `module_inception` in two `tests.rs` files: flattened the inner
  `mod tests`, and gated active_theme's `mod tests;` behind cfg(test).
- Targeted allows, each with a reason: benches (`unwrap_used` in setup),
  the oasis-ffi test module (one SAFETY note for its ~130 unsafe calls
  into the crate's own exports, plus the null checks it relies on),
  `clone_on_copy` in the test that exists to exercise Clone, and
  `ptr_arg` where `ListView<String>`'s callback type forces `&String`.

CI (main + PR), the nightly advisory step, the pre-commit hooks,
run-ci.sh and the docs now all use
`cargo clippy --workspace --all-targets -- -D warnings`.


* fix(browser): drop stale node references when the document is replaced

NodeIds are arena indices into the current Document, but hover_node,
focused_node, tab_order and the transition engine survived page loads.
Hovering a node deep in a large page and then loading a smaller one
(e.g. click-launching the Browser app from the dashboard) left
hover_node past the end of the new DOM, and the next cursor move
panicked with "index out of bounds" in restyle_hover_affected. Stale
ids that stayed in range silently applied :hover/:focus and running
transitions to unrelated nodes of the new page.

Clear that state in reset_for_navigation, in load_html before the
cascade (which reads hover_node), and on the reader-mode swap. The
:hover/:focus ancestor walks now use checked lookups as a backstop.

Also gate the e2e harness's wait_for on the javascript feature; only
the JS e2e tests call it, so crate-only builds warned on dead code.


* fix(browser): scroll pages that overflow a viewport-height body; speed up style and layout

Wikipedia sets `html, body { height: 100% }`, which pins the root box to
the viewport while the article overflows it. The scrollable height came
from the root box alone (max scroll 0), paint culling dropped the whole
overflowing subtree with its off-screen ancestor, and hit testing stopped
at the root's bounds. Measure the document from the root's visible
overflow (`LayoutBox::scrollable_height`), and only cull / reject a box on
its own bounds when it clips its contents or has no children.

Loading also froze the shell for seconds per step on Wikipedia:
- ComputedStyle deep-copied its custom-property map (~200 design
  tokens, inherited by every element) on each clone. Share it via Arc
  copy-on-write: full relayout 1.7s -> 40ms (release).
- ::before/::after resolution scanned every rule of every sheet twice
  per element. Index pseudo-element rules by subject like ordinary
  rules: re-cascade ~1s -> ~150ms.


* fix(browser): make oasis-browser clippy-clean with --all-features

`--all-features` turns on `psp` and `parallel-style`, which CI doesn't
build, and four lints had crept in there:
- `budget_ms` is unused on PSP (time budgeting is skipped there).
- `cascade_yield_fn` is only called by the sequential cascade.
- The parallel cascade path used a needless `return`.
- A nested redirect `if` in the PSP HTTP loader could be collapsed.


* fix(shader): gate the golden-checksum hash helper to x86_64 like its only test

`fnv1a` is used only by `golden_output_checksums`, which is gated to
`target_arch = "x86_64"` (the goldens depend on the platform libm).
On the aarch64 self-hosted runner the helper was dead code, failing
`cargo clippy --workspace --all-targets -- -D warnings`.


---------

+++ crates/oasis-wm/src/manager/mod.rs
@@ -1575,8 +1575,8 @@ mod tests {
             let config = WindowConfig {
                 id: format!("w{i}"),
                 title: format!("W{i}"),
-                x: Some(i as i32 * 10),
-                y: Some(i as i32 * 10),
+                x: Some(i * 10),
+                y: Some(i * 10),
                 width: 100,
                 height: 80,
                 window_type: WindowType::AppWindow,
